Rename the chain matcher
