+++ robber/__init__.py
@@ -1,16 +1,4 @@
-# TODO: fix me
 from expect import expect
 from failure_message import failure_message
-
-import matchers.equal
-import matchers.identical
-import matchers.boolean
-import matchers.instanceof
-import matchers.regexp
-import matchers.respond_to
-import matchers.truthy
-import matchers.length
-import matchers.types
-import matchers.numbers
-
 from bad_expectation import BadExpectation
+from matchers import *
