+++ robber/matchers/__init__.py
@@ -1,0 +1,5 @@
+__all__ = [
+    'equal', 'identical', 'boolean', 'instanceof',
+    'regexp', 'respond_to', 'truthy', 'length',
+    'types', 'numbers'
+]
