+++ tests/test_expect.py
@@ -1,9 +1,8 @@
 import unittest
 from robber import expect, BadExpectation
 from fixtures import TestMatcher
-import matcher
+import chain_matcher
 
-# TODO: move to fixtures
 class TestExpectation(unittest.TestCase):
     def test_register_a_matcher(self):
         expect.register('test_matcher', TestMatcher)
